Extract TagViewModelBase
